+++ src/main/java/org/kdb/inside/brains/core/credentials/CredentialPluginsPanel.java
@@ -4,7 +4,6 @@ import com.intellij.openapi.fileChooser.FileChooserDescriptor;
 import com.intellij.openapi.fileChooser.FileChooserDialog;
 import com.intellij.openapi.fileChooser.FileChooserFactory;
 import com.intellij.openapi.ui.Messages;
-import com.intellij.openapi.vfs.VfsUtilCore;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.TableUtil;
 import com.intellij.ui.ToolbarDecorator;
@@ -17,20 +16,19 @@ import org.jetbrains.annotations.NotNull;
 import javax.swing.*;
 import javax.swing.table.AbstractTableModel;
 import java.awt.*;
-import java.net.URL;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 public class CredentialPluginsPanel extends JPanel implements ComponentWithEmptyText {
-    private static final String[] COLUMN_NAMES = {"Name", "Version", "Description"};
-    private final CredentialService credentialService;
-    private final List<CredentialPlugin> plugins = new ArrayList<>();
     private JBTable myTable;
     private AbstractTableModel myTableModel;
 
+    private static final String[] COLUMN_NAMES = {"Name", "Version", "Description"};
+    private final List<PluginRef> plugins = new ArrayList<>();
+
     public CredentialPluginsPanel() {
-        credentialService = CredentialService.getInstance();
         initTable();
         initPanel();
     }
@@ -59,11 +57,11 @@ public class CredentialPluginsPanel extends JPanel implements ComponentWithEmpty
 
             @Override
             public Object getValueAt(int rowIndex, int columnIndex) {
-                final CredentialPlugin plugin = plugins.get(rowIndex);
+                final PluginDescriptor plugin = plugins.get(rowIndex).descriptor;
                 return switch (columnIndex) {
-                    case 0 -> plugin.getName();
-                    case 1 -> plugin.getVersion();
-                    case 2 -> plugin.getDescription();
+                    case 0 -> plugin.name();
+                    case 1 -> plugin.version();
+                    case 2 -> plugin.description();
                     default -> "";
                 };
             }
@@ -99,8 +97,7 @@ public class CredentialPluginsPanel extends JPanel implements ComponentWithEmpty
         Arrays.sort(selected);
 
         for (int i = selected.length - 1; i >= 0; i--) {
-            int idx = selected[i];
-            plugins.remove(idx);
+            plugins.remove(selected[i]);
         }
 
         myTableModel.fireTableDataChanged();
@@ -140,20 +137,20 @@ public class CredentialPluginsPanel extends JPanel implements ComponentWithEmpty
         }
 
         final VirtualFile virtualFile = choose[0];
-        final URL url = VfsUtilCore.convertToURL(virtualFile.getUrl());
-        if (url == null) {
+        if (!virtualFile.exists() || !virtualFile.isValid()) {
             return;
         }
+        final Path resource = virtualFile.toNioPath();
 
         try {
-            final CredentialPlugin plugin = credentialService.verifyPlugin(url);
-
-            final int index = findPlugin(plugin.getId());
+            final PluginDescriptor d = CredentialService.verifyPlugin(resource);
+            final PluginRef ref = new PluginRef(d, resource);
+            final int index = findPlugin(d.id());
             if (index < 0) {
-                plugins.add(plugin);
+                plugins.add(ref);
                 myTableModel.fireTableRowsInserted(plugins.size() - 1, plugins.size() - 1);
             } else {
-                plugins.set(index, plugin);
+                plugins.set(index, ref);
                 myTableModel.fireTableRowsUpdated(index, index);
             }
         } catch (Exception ex) {
@@ -163,8 +160,8 @@ public class CredentialPluginsPanel extends JPanel implements ComponentWithEmpty
 
     private int findPlugin(String id) {
         int i = 0;
-        for (CredentialPlugin plugin : plugins) {
-            if (id.equals(plugin.getId())) {
+        for (PluginRef plugin : plugins) {
+            if (plugin.is(id)) {
                 return i;
             }
             i++;
@@ -172,13 +169,23 @@ public class CredentialPluginsPanel extends JPanel implements ComponentWithEmpty
         return -1;
     }
 
-    public List<CredentialPlugin> getCredentialPlugins() {
-        return new ArrayList<>(plugins);
+    public List<Path> getPluginResources() {
+        return plugins.stream().map(PluginRef::resource).toList();
     }
 
-    public void setCredentialPlugins(List<CredentialPlugin> credentialPlugins) {
+    public void updateCredentialPlugins(List<CredentialPlugin> credentialPlugins) {
         plugins.clear();
-        plugins.addAll(credentialPlugins);
+        credentialPlugins.stream().map(PluginRef::new).forEachOrdered(plugins::add);
         myTableModel.fireTableDataChanged();
+    }
+
+    private record PluginRef(PluginDescriptor descriptor, Path resource) {
+        private PluginRef(CredentialPlugin plugin) {
+            this(plugin.getDescriptor(), plugin.getResource());
+        }
+
+        private boolean is(String id) {
+            return id.equals(descriptor.id());
+        }
     }
 }